+++ Sources/App/Helper/Download/S3List.swift
@@ -1,7 +1,6 @@
 import AsyncHTTPClient
 import Foundation
-import Vapor
-
+import Logging
 
 enum S3List {
     /// https://docs.aws.amazon.com/AmazonS3/latest/API/API_ListObjectsV2.html
