+++ Sources/App/Helper/File/OmFileSystemS3.swift
@@ -1,9 +1,5 @@
 import AsyncHTTPClient
-import Foundation
 import Logging
-import NIO
-import NIOConcurrencyHelpers
-import Vapor
 import OmFileFormat
 
 enum OmFileSystemS3Error: Error {
